Fixing the GUI runtime error
This commit changes the way irradiance is set up as a matrix in fit_all_solar. The mismatch between irradiance and data matrices was giving us the error. Irradiance was set up as only the dark days, but then was fed indexes of the dark days when assigning values, giving it a weird shape. I believe the intention was to only feed in irradiance data on bright days, so I've treated it as such

(Also I've removed the things that should be spamming the console with values for the most part)